requirements.txt
AWSIoTPythonSDK==1.4.6
azure-iothub-device-client==1.4.6
azure-iothub-service-client==1.4.6
paho-mqtt==1.4.0
